requirements.txt
google-api-python-client
youtube-transcript-api
python-dotenv
pyyaml
google-genai
